Moved site_persistence files within the project. Generated a new `config` directory with nested directories for `eproms` and `gil`. Moved implementation modules `config.py` and `site_persistence.py` within. Updated docs to fit new model. Replaced PERSISTENCE_FILE with PERSISTENCE_DIR (and corrected value) as we plan to move forward with any number of persistence files rather than a single, massive, global one per project.
