+++ app.json
@@ -11,9 +11,9 @@
       "description": "Fully qualified domain name of system, excluding protocol (HTTP/S). See App Name",
       "value": "HEROKU_APP_NAME.herokuapp.com"
     },
-    "PERSISTENCE_FILE": {
-      "description": "Site-specific configuration file, defaults to TrueNTH",
-      "value": "https://raw.githubusercontent.com/uwcirg/TrueNTH-USA-site-config/develop/site_persistence_file.json"
+    "PERSISTENCE_DIR": {
+      "description": "Site-specific configuration directory, defaults to gil",
+      "value": "gil"
     }
   },
   "addons": [
