+++ docs/source/configuration.rst
@@ -32,19 +32,20 @@ class.  Most are self explanitory or include inline comments for clarification.
 Of special note, the one used to control which set of values are pulled in
 by `Site Persistence`_.
 
-PERSISTENCE_FILE::
+PERSISTENCE_DIR::
 
-    See also Site Persistence, this controls which persistence file the
-    `FLASK_APP=manage.py flask seed` command uses to load persistence data and build
-    the `site.cfg` file.
+    See also Site Persistence, this controls which persistence directory the
+    `FLASK_APP=manage.py flask sync` command uses to load persistence data
+    and build the `site.cfg` file.  The value is relative to the
+    `portal/config` directory.
 
     For TrueNTH:
 
-        PERSISTENCE_FILE='https://raw.githubusercontent.com/uwcirg/TrueNTH-USA-site-config/master/site_persistence_file.json'
+        PERSISTENCE_DIR='gil'
 
     For ePROMs:
 
-        PERSISTENCE_FILE='https://raw.githubusercontent.com/uwcirg/ePROMs-site-config/master/site_persistence_file.json'
+        PERSISTENCE_DIR='eproms'
 
 Site Persistence
 ================
@@ -64,33 +65,27 @@ will include codified business rules and required data to support them.
 
 Database tables included:
 
-- Organizations
-- Interventions
 - AccessStrategies
 - AppText
+- CommunicationRequest
+- Interventions
+- Organizations
+- Questionnaires
+- QuestionnaireBanks
+- ScheduledJobs
 
-Both importing and exporting use the value of ``PERSISTENCE_FILE``.
-Its value is initially looked up in the configuration file(s) and subsequently
-as an environment variable - with the first well defined value taking
-precedence.
+Both importing and exporting use the value of ``PERSISTENCE_DIR``.
+Its value is initially looked for as an environment variable, and if not
+found, the configuration value of 'GIL' is used.  (With 'GIL' set, the `gil`
+configuration directory is used, othersise, `eproms`).
 
 Export
 ------
-A site persistence file can be generated by setting the config value
-`PERSISTENCE_FILE` to any writable file on the local filesystem, and executing:
+Site persistence files can be generated in the ``PERSISTENCE_DIR``.  See
+above for correct setting.  To generate persistence files from current
+database values, execute::
 
 ```FLASK_APP=manage.py flask export_site```
-
-The resulting file can be added to a repository for ease of use by other users.
-Existing repositories include:
-
-- `TrueNTH-site-config <https://github.com/uwcirg/TrueNTH-USA-site-config>`_
-- `ePROMs-site-config <https://github.com/uwcirg/ePROMs-site-config>`_
-
-.. note::
-
-   For exports, **PERSISTENCE_FILE** must point to a writable file on the local
-   filesystem.
 
 Import
 ------
