+++ docs/source/portal.rst
@@ -11,13 +11,7 @@ Portal
     .. automodule:: portal.celery_worker
         :members:
 
-.. automodule:: portal.config
-    :members:
-
 .. automodule:: portal.extensions
-    :members:
-
-.. automodule:: portal.site_persistence
     :members:
 
 .. automodule:: portal.system_uri
