+++ docs/source/portal_config.rst
@@ -1,0 +1,7 @@
+Portal.Config
+=============
+.. automodule:: portal.config.config
+    :members:
+
+.. automodule:: portal.config.site_persistence
+    :members:
