+++ manage.py
@@ -20,7 +20,7 @@ from portal.models.relationship import add_static_relationships
 from portal.models.role import add_static_roles
 from portal.models.user import permanently_delete_user, flag_test
 from portal.models.user_consent import db_maintenance
-from portal.site_persistence import SitePersistence
+from config.site_persistence import SitePersistence
 
 app = create_app()
