+++ tests/__init__.py
@@ -13,7 +13,7 @@ from flask_webtest import SessionScope
 from sqlalchemy.exc import IntegrityError
 
 from portal.factories.app import create_app
-from portal.config import TestConfig
+from portal.config.config import TestConfig
 from portal.extensions import db
 from portal.models.assessment_status import invalidate_assessment_status_cache
 from portal.models.audit import Audit
