+++ tox.ini
@@ -8,7 +8,7 @@ commands =
 
 [testenv]
 deps = -rrequirements.txt
-passenv = TRAVIS* PG* SQLALCHEMY_DATABASE_TEST_URI PERSISTENCE_FILE FLASK_APP
+passenv = TRAVIS* PG* SQLALCHEMY_DATABASE_TEST_URI PERSISTENCE_DIR FLASK_APP
 whitelist_externals = /bin/bash
 commands =
     {[base]commands}
